Added option for -- at end of prototype line to place remaining arguments into REMAINING
Added option that if "--" appears at end of prototype line that if
"--" is encountered on input line remaining arguments go into
requoted character variable REMAINING instead of as entries in
UNNAMED. Operation remains the same unless the prototype ends in
"--". Arguments unassociated with a keyword continue to go into
UNNAMED until a "--" is explicitly encounted on input.