Special mode when prototype name ends in :: forces use of next argument as a value
To support a special option on fpm(1) that takes an argument starting
with a dash made option triggered by -NAME:: in prototype that forces
next command line argument to be used as a value.

Also added another option that uses -h -v and -u for short names for
--help --version and --usage if those letters are not used by the user.

Need to restructure logic to make it a bit cleaner after adding such odd
options.